requirements.txt
Flask==2.3.3
flasgger==0.9.7.1
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.0

libversion @ git+https://github.com/remla25-team3/lib-version@v1.2.0
lib-ml-remla-team3 @ git+https://github.com/remla25-team3/lib-ml@v0.5.0

.   # Import packages inside this project